LayoutTests/fast/text/font-download-font-face-src-list.html

<!DOCTYPE html><!-- webkit-test-runner [ dumpResourceLoadCallbacks=true ] -->
<html>
<head>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
}
if (window.internals) {
    internals.invalidateFontCache();
    internals.clearMemoryCache();
}
</script>
<style>
@font-face {
    font-family: "WebFont";
    src: url("../../resources/Ahem_CJK.ttf") format("truetype"), url("../../resources/Ahem.otf") format("opentype");
}
</style>
</head>
<body>
This test makes sure that unnecessary fonts aren't downloaded. The test fails if Ahem.otf is downloaded.
<div style="font-size: 100px; font-family: 'WebFont';">&#x6A2A;</div>
</body>
</html>
